Add files via upload
The Weather App is a real-time project built using JavaScript, CSS, Express, and Handlebars. It serves as an engaging platform for basic weather forecasting. Users can check the current weather conditions and assess whether the day is likely to be sunny or cloudy. By entering the name of any city worldwide, they can retrieve real-time climate data through the app.

To run the APP you must run it on Localhost 8000